mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.50.0",
    "azure-identity>=1.21.0",
    "azure-mgmt-compute>=34.1.0",
    "azure-mgmt-costmanagement>=4.0.1",
    "azure-mgmt-resource>=23.3.0",
    "google-api-python-client>=2.169.0",
    "google-cloud-billing>=1.16.2",
    "google-cloud-billing-budgets>=1.17.1",
    "mcp>=1.6.0",
    "ollama>=0.4.8",
    "openai>=1.76.0",
    "proxmoxmanager>=1.0.4",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
